Gjør om RegelverkFørFebruar2025AndelGeneratorGammel og OppdaterAndelerMedEndretUtbetalingService til object.
