+++ src/main/kotlin/no/nav/familie/ks/sak/kjerne/beregning/endretUtbetaling/OppdaterAndelerMedEndretUtbetalingService.kt
@@ -10,12 +10,10 @@ import no.nav.familie.ks.sak.kjerne.beregning.AndelTilkjentYtelseMedEndreteUtbet
 import no.nav.familie.ks.sak.kjerne.beregning.EndretUtbetalingAndelMedAndelerTilkjentYtelse
 import no.nav.familie.ks.sak.kjerne.beregning.domene.AndelTilkjentYtelse
 import no.nav.familie.ks.sak.kjerne.beregning.medEndring
-import org.springframework.stereotype.Service
 import java.math.BigDecimal
 import java.time.YearMonth
 
-@Service
-class OppdaterAndelerMedEndretUtbetalingService {
+object OppdaterAndelerMedEndretUtbetalingService {
     fun oppdaterTilkjentYtelseMedEndretUtbetalingAndeler(
         andelTilkjentYtelserUtenEndringer: List<AndelTilkjentYtelse>,
         endretUtbetalingAndeler: List<EndretUtbetalingAndelMedAndelerTilkjentYtelse>,
