+++ src/test/enhetstester/kotlin/no/nav/familie/ks/sak/kjerne/beregning/TilkjentYtelseServiceTest.kt
@@ -2,12 +2,9 @@ package no.nav.familie.ks.sak.kjerne.beregning
 
 import no.nav.familie.ks.sak.common.util.NullablePeriode
 import no.nav.familie.ks.sak.common.util.førsteDagIInneværendeMåned
-import no.nav.familie.ks.sak.common.util.nesteMåned
 import no.nav.familie.ks.sak.common.util.sisteDagIMåned
 import no.nav.familie.ks.sak.common.util.toYearMonth
-import no.nav.familie.ks.sak.data.lagAndelTilkjentYtelse
 import no.nav.familie.ks.sak.data.lagBehandling
-import no.nav.familie.ks.sak.data.lagEndretUtbetalingAndel
 import no.nav.familie.ks.sak.data.lagPerson
 import no.nav.familie.ks.sak.data.lagPersonopplysningGrunnlag
 import no.nav.familie.ks.sak.data.lagVilkårResultaterForBarn
@@ -24,8 +21,6 @@ import no.nav.familie.ks.sak.kjerne.beregning.domene.TilkjentYtelse
 import no.nav.familie.ks.sak.kjerne.beregning.domene.YtelseType
 import no.nav.familie.ks.sak.kjerne.beregning.domene.maksBeløp
 import no.nav.familie.ks.sak.kjerne.beregning.domene.prosent
-import no.nav.familie.ks.sak.kjerne.beregning.endretUtbetaling.OppdaterAndelerMedEndretUtbetalingService
-import no.nav.familie.ks.sak.kjerne.beregning.regelverkFørFebruar2025.gammel.RegelverkFørFebruar2025AndelGeneratorGammel
 import no.nav.familie.ks.sak.kjerne.personopplysninggrunnlag.domene.PersonType
 import org.junit.jupiter.api.Assertions.assertEquals
 import org.junit.jupiter.api.Assertions.assertTrue
@@ -33,9 +28,8 @@ import org.junit.jupiter.api.BeforeEach
 import org.junit.jupiter.api.Test
 import java.math.BigDecimal
 import java.time.LocalDate
-import java.time.YearMonth
 
-internal class TilkjentYtelseUtilsTest {
+internal class TilkjentYtelseServiceTest {
     private val søker = randomAktør()
 
     private val barn1 = randomAktør("01012112345")
@@ -56,13 +50,7 @@ internal class TilkjentYtelseUtilsTest {
 
     private lateinit var vilkårsvurdering: Vilkårsvurdering
 
-    private val oppdaterAndelerMedEndretUtbetalingService = OppdaterAndelerMedEndretUtbetalingService()
-    private val regelverkFørFebruar2025AndelGeneratorGammel = RegelverkFørFebruar2025AndelGeneratorGammel()
-    private val tilkjentYtelseService =
-        TilkjentYtelseService(
-            regelverkFørFebruar2025AndelGeneratorGammel = regelverkFørFebruar2025AndelGeneratorGammel,
-            oppdaterAndelerMedEndretUtbetalingService = oppdaterAndelerMedEndretUtbetalingService,
-        )
+    private val tilkjentYtelseService = TilkjentYtelseService()
 
     @BeforeEach
     fun init() {
@@ -522,112 +510,6 @@ internal class TilkjentYtelseUtilsTest {
             periodeFom = andrePeriodeFom.førsteDagIInneværendeMåned(),
             periodeTom = barnFødselsdato.plusYears(2).minusMonths(1).sisteDagIMåned(),
         )
-    }
-
-    @Test
-    fun `oppdaterTilkjentYtelseMedEndretUtbetalingAndeler - endret utbetalingsandel skal overstyre andel`() {
-        val fom = YearMonth.of(2018, 1)
-        val tom = YearMonth.of(2019, 1)
-
-        val utbetalingsandeler =
-            listOf(
-                lagAndelTilkjentYtelse(
-                    stønadFom = fom,
-                    stønadTom = tom,
-                    aktør = søker,
-                    behandling = behandling,
-                ),
-            )
-
-        val endretProsent = BigDecimal.ZERO
-
-        val endretUtbetalingAndel =
-            lagEndretUtbetalingAndel(
-                behandlingId = behandling.id,
-                person = søkerPerson,
-                periodeFom = fom,
-                periodeTom = tom,
-                prosent = BigDecimal.ZERO,
-            )
-
-        val endretUtbetalingAndelMedAndelerTilkjentYtelse =
-            EndretUtbetalingAndelMedAndelerTilkjentYtelse(endretUtbetalingAndel, utbetalingsandeler)
-
-        val andelerTilkjentYtelse =
-            oppdaterAndelerMedEndretUtbetalingService.oppdaterTilkjentYtelseMedEndretUtbetalingAndeler(
-                utbetalingsandeler,
-                listOf(endretUtbetalingAndelMedAndelerTilkjentYtelse),
-            )
-
-        assertEquals(1, andelerTilkjentYtelse.size)
-        assertEquals(endretProsent, andelerTilkjentYtelse.single().prosent)
-        assertEquals(1, andelerTilkjentYtelse.single().endreteUtbetalinger.size)
-    }
-
-    @Test
-    fun `oppdaterTilkjentYtelseMedEndretUtbetalingAndeler - endret utbetalingsandel koble endrede andeler til riktig endret utbetalingandel`() {
-        val fom1 = YearMonth.of(2018, 1)
-        val tom1 = YearMonth.of(2018, 11)
-
-        val fom2 = YearMonth.of(2019, 1)
-        val tom2 = YearMonth.of(2019, 11)
-
-        val utbetalingsandeler =
-            listOf(
-                lagAndelTilkjentYtelse(
-                    stønadFom = fom1,
-                    stønadTom = tom1,
-                    aktør = søker,
-                    behandling = behandling,
-                ),
-                lagAndelTilkjentYtelse(
-                    stønadFom = fom2,
-                    stønadTom = tom2,
-                    aktør = søker,
-                    behandling = behandling,
-                ),
-            )
-
-        val endretProsent = BigDecimal.ZERO
-
-        val endretUtbetalingAndel =
-            lagEndretUtbetalingAndel(
-                behandlingId = behandling.id,
-                person = søkerPerson,
-                periodeFom = fom1,
-                periodeTom = tom2,
-                prosent = BigDecimal.ZERO,
-            )
-
-        val endretUtbetalingAndelMedAndelerTilkjentYtelse1 =
-            EndretUtbetalingAndelMedAndelerTilkjentYtelse(endretUtbetalingAndel, utbetalingsandeler)
-
-        val endretUtbetalingAndel2 =
-            lagEndretUtbetalingAndel(
-                behandlingId = behandling.id,
-                person = søkerPerson,
-                periodeFom = tom2.nesteMåned(),
-                prosent = endretProsent,
-            )
-
-        val endretUtbetalingAndelMedAndelerTilkjentYtelse2 =
-            EndretUtbetalingAndelMedAndelerTilkjentYtelse(endretUtbetalingAndel2, utbetalingsandeler)
-
-        val andelerTilkjentYtelse =
-            oppdaterAndelerMedEndretUtbetalingService.oppdaterTilkjentYtelseMedEndretUtbetalingAndeler(
-                utbetalingsandeler,
-                listOf(endretUtbetalingAndelMedAndelerTilkjentYtelse1, endretUtbetalingAndelMedAndelerTilkjentYtelse2),
-            )
-
-        assertEquals(2, andelerTilkjentYtelse.size)
-        andelerTilkjentYtelse.forEach { assertEquals(endretProsent, it.prosent) }
-        andelerTilkjentYtelse.forEach { assertEquals(1, it.endreteUtbetalinger.size) }
-        andelerTilkjentYtelse.forEach {
-            assertEquals(
-                endretUtbetalingAndel.id,
-                it.endreteUtbetalinger.single().id,
-            )
-        }
     }
 
     private fun assertTilkjentYtelse(
