+++ src/test/enhetstester/kotlin/no/nav/familie/ks/sak/kjerne/eøs/differanseberegning/TilkjentYtelseDifferanseberegningTest.kt
@@ -11,7 +11,6 @@ import no.nav.familie.ks.sak.kjerne.behandling.steg.vilkårsvurdering.domene.Vil
 import no.nav.familie.ks.sak.kjerne.behandling.steg.vilkårsvurdering.domene.Vilkår.LOVLIG_OPPHOLD
 import no.nav.familie.ks.sak.kjerne.behandling.steg.vilkårsvurdering.domene.Vilkår.MEDLEMSKAP
 import no.nav.familie.ks.sak.kjerne.behandling.steg.vilkårsvurdering.domene.Vilkår.MEDLEMSKAP_ANNEN_FORELDER
-import no.nav.familie.ks.sak.kjerne.beregning.endretUtbetaling.OppdaterAndelerMedEndretUtbetalingService
 import no.nav.familie.ks.sak.kjerne.eøs.util.DeltBostedBuilder
 import no.nav.familie.ks.sak.kjerne.eøs.util.TilkjentYtelseBuilder
 import no.nav.familie.ks.sak.kjerne.eøs.util.UtenlandskPeriodebeløpBuilder
@@ -31,8 +30,6 @@ import java.time.LocalDate
  */
 class TilkjentYtelseDifferanseberegningTest {
     private fun Int.jan(år: Int): LocalDate = LocalDate.of(år, 1, this)
-
-    private val oppdaterAndelerMedEndretUtbetalingService = OppdaterAndelerMedEndretUtbetalingService()
 
     @Test
     fun `skal gjøre differanseberegning på en tilkjent ytelse med endringsperioder`() {
@@ -65,7 +62,7 @@ class TilkjentYtelseDifferanseberegningTest {
 
         DeltBostedBuilder(startMåned, tilkjentYtelse)
             .medDeltBosted(" //////00000000001111>", barn1, barn2)
-            .oppdaterTilkjentYtelse(oppdaterAndelerMedEndretUtbetalingService)
+            .oppdaterTilkjentYtelse()
 
         val forventetTilkjentYtelseMedDelt =
             TilkjentYtelseBuilder(startMåned, behandling)
